Udpate cron, improve parseComic
CRON nows run Monday, Wednesday and Friday every half hour beween 12 and
18.
ParseComic now parses comics as long as there is now error, it can be
useful if many comics are published on the same day or if the bot missed
one previous comic (downtime or any other reason).
Note that this will not work if a ID is missing on XKCD, but it should not be
the case.